requirements.txt
asgiref==3.8.1
cryptography==45.0.5
Django==4.2.25
django-cors-headers==4.4.0
django-extensions==3.2.3
django-multiselectfield==0.1.13
django-phonenumber-field==7.3.0
djangorestframework==3.15.2
drf_base64==2.0
drfpasswordless==1.5.9
Faker==13.12.1
pillow==10.4.0
phonenumbers==8.10.20
PyJWT==2.9.0
python-dateutil==2.8.2
python-dotenv==1.0.1
reportlab==3.6.13
six==1.12.0
sqlparse==0.5.0
typing_extensions==4.11.0

# dev requirements
pygraphviz==1.11
pre_commit==3.5.0

# production requirements
gunicorn==23.0.0
